Day4

+++ backend/controllers/GoodsCategoryController.php
@@ -46,20 +46,31 @@ class GoodsCategoryController extends \yii\web\Controller
     }
     //修改
     public function actionEdit($id){
+        //节点不能修改上级分类是自己或是自己的子孙节点
         $model=GoodsCategory::findOne(['id'=>$id]);
         $request=\Yii::$app->request;
         if ($request->isPost){
+            $ord_parent_id=$model->parent_id;
             $model->load($request->post());
             if ($model->validate()){
-                if ($model->parent_id){
-                    //创建子孙节点
-                    $parent=GoodsCategory::findOne(['id'=>$model->parent_id]);
-                    $model->prependTo($parent);
+                if ($model->parent_id==$model->getOldAttribute('parent_id')){
+                    $model->save();
                 }else{
-                    //创建根节点
-                    $model->makeRoot();
+                    if ($model->parent_id){
+                        //创建子孙节点
+                        $parent=GoodsCategory::findOne(['id'=>$model->parent_id]);
+                        $model->prependTo($parent);
+                    }else{
+                        //顶级分类不能修改为顶级分类,即旧节点父id为0,新节点父id也为0
+                        if ($model->getOldAttribute('parent_id')==0){//或者$ord_parent_id==0
+                            $model->save();
+                        }else{
+                            //创建根节点
+                            $model->makeRoot();
+                        }
+                    }
+                    $model->save();
                 }
-                $model->save();
                 \Yii::$app->session->setFlash('success','修改成功');
                 return $this->redirect(['goods-category/index']);
             }else{
@@ -73,13 +84,18 @@ class GoodsCategoryController extends \yii\web\Controller
     //删除
     public function actionDelete($id){
         $model=GoodsCategory::findOne(['id'=>$id]);
-        if ($model->parent_id){
-            $model->delete();
-            \Yii::$app->session->setFlash('success','删除成功');
-            return $this->redirect(['goods-category/index']);
+        $child=GoodsCategory::find()->where(['parent_id'=>$id]);
+        if ($child==null){
+            if ($model->parent_id){
+                $model->delete();
+                \Yii::$app->session->setFlash('success','删除成功');
+            }else{
+                \Yii::$app->session->setFlash('danger','该分类为根分类,不能删除');
+            }
         }else{
+            //$model->deleteWithChildren();删除当前节点和子孙节点
             \Yii::$app->session->setFlash('danger','该分类下有子分类');
-            return $this->redirect(['goods-category/index']);
         }
+        return $this->redirect(['goods-category/index']);
     }
 }
